+++ lensless/recon/trainable_recon.py
@@ -5,8 +5,10 @@
 # Yohann PERRON [yohann.perron@gmail.com]
 # #############################################################################
 
-import abc
+import pathlib as plib
+from matplotlib import pyplot as plt
 from lensless.recon.recon import ReconstructionAlgorithm
+from lensless.utils.plot import plot_image
 
 try:
     import torch
@@ -24,7 +26,6 @@ class TrainableReconstructionAlgorithm(ReconstructionAlgorithm, torch.nn.Module)
     * ``_update``: updating state variables at each iterations.
     * ``reset``: reset state variables.
     * ``_form_image``: any pre-processing that needs to be done in order to view the image estimate, e.g. reshaping or clipping.
-    * ``batch_call``: method for performing iterative reconstruction on a batch of images.
 
     One advantage of deriving from this abstract class is that functionality for
     iterating, saving, and visualization is already implemented, namely in the
@@ -155,7 +156,15 @@ class TrainableReconstructionAlgorithm(ReconstructionAlgorithm, torch.nn.Module)
         return image_est
 
     def apply(
-        self, disp_iter=10, plot_pause=0.2, plot=True, save=False, gamma=None, ax=None, reset=True
+        self,
+        disp_iter=10,
+        plot_pause=0.2,
+        plot=True,
+        save=False,
+        gamma=None,
+        ax=None,
+        reset=True,
+        output_intermediate=False,
     ):
         """
         Method for performing iterative reconstruction. Contrary to non-trainable reconstruction
@@ -180,6 +189,8 @@ class TrainableReconstructionAlgorithm(ReconstructionAlgorithm, torch.nn.Module)
             Gamma correction factor to apply for plots. Default is None.
         ax : :py:class:`~matplotlib.axes.Axes`, optional
             `Axes` object to fill for plotting/saving, default is to create one.
+        output_intermediate : bool, optional
+            Whether to output intermediate reconstructions after preprocessing and before postprocessing.
 
         Returns
         -------
@@ -190,8 +201,11 @@ class TrainableReconstructionAlgorithm(ReconstructionAlgorithm, torch.nn.Module)
             returning if `plot` or `save` is True.
 
         """
+        pre_processed_image = None
         if self.pre_process is not None:
             self._data = self.pre_process(self._data, self.pre_process_param)
+            if output_intermediate:
+                pre_processed_image = self._data[0, ...].clone()
 
         im = super(TrainableReconstructionAlgorithm, self).apply(
             n_iter=self._n_iter,
@@ -203,6 +217,30 @@ class TrainableReconstructionAlgorithm(ReconstructionAlgorithm, torch.nn.Module)
             ax=ax,
             reset=reset,
         )
+
+        # remove plot if returned
+        if plot:
+            im, _ = im
+
+        # post process data
+        pre_post_process_image = None
         if self.post_process is not None:
+            # apply post process
+            if output_intermediate:
+                pre_post_process_image = im.clone()
             im = self.post_process(im, self.post_process_param)
-        return im
+
+        if plot:
+            ax = plot_image(self._get_numpy_data(im[0]), ax=ax, gamma=gamma)
+            ax.set_title(
+                "Final reconstruction after {} iterations and post process".format(self._n_iter)
+            )
+            if save:
+                plt.savefig(plib.Path(save) / "final.png")
+
+        if output_intermediate:
+            return im, pre_processed_image, pre_post_process_image
+        elif plot:
+            return im, ax
+        else:
+            return im